views de cadastro movidas para accounts/views
